bug-triage-agent/requirements.txt
# Bug Triage AI Agent - Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
pymongo==4.6.0

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Logging
python-json-logger==2.0.7

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.1

